refactor(admin): move user and tariff stores to runes
